+++ src/page/Order/hooks/useOrderForm.ts
@@ -1,43 +1,18 @@
-import { useNavigate, useParams } from 'react-router-dom';
-import { useUserInfo } from '@/contexts/UserInfoContext';
+import { useParams } from 'react-router-dom';
 import { toast } from 'react-toastify';
-import { ROUTE_PATH } from '@/routes/routePath';
-import { useForm } from 'react-hook-form';
 import type { OrderInfoValues } from '@/types';
 import useProductSummary from './useProductSummary';
-import { useMutation } from '@tanstack/react-query';
-import { requests } from '@/api/requests';
-import axios from 'axios';
+import useSubmitOrder from './useSubmitOrder';
+import showOrderSuccessAlert from '../utils/showOrderSuccessAlert';
+import useInitOrderForm from './useInitOrderForm';
 
 const useOrderForm = () => {
   const { id } = useParams<{ id: string }>();
-  const navigate = useNavigate();
   const productSummaryData = useProductSummary(id as string);
-
-  const { userInfo } = useUserInfo();
-  const orderForm = useForm<OrderInfoValues>({
-    defaultValues: { message: '축하해요.', name: userInfo.name, receiverInfos: [] },
-  });
+  const orderForm = useInitOrderForm();
   const orderData = orderForm.getValues();
   const price = orderData.receiverInfos.length * (productSummaryData?.price || 0);
-
-  const { mutate: order } = useMutation({
-    mutationFn: requests.fetchOrder,
-    onSuccess: () => {
-      alert(`
-      주문이 완료되었습니다.
-      상품명: ${productSummaryData?.name}
-      구매 수량: ${orderData.receiverInfos.length}
-      발신자 이름: ${orderData.name}
-      메시지: ${orderData.message}`);
-      navigate(ROUTE_PATH.HOME);
-    },
-    onError: error => {
-      if (axios.isAxiosError(error) && error.response?.status === 401) {
-        navigate(ROUTE_PATH.LOGIN);
-      }
-    },
-  });
+  const { mutate: order } = useSubmitOrder();
 
   const onSubmit = async (formData: OrderInfoValues) => {
     if (!id) return;
@@ -46,6 +21,7 @@ const useOrderForm = () => {
       return;
     }
     order({ orderData: formData, id });
+    showOrderSuccessAlert({ productSummaryData, formData });
   };
 
   return { orderForm, onSubmit, price, productSummaryData };
